Move SanitizedAttribute to a better namespace
