+++ src/UserAdmin/src/Smart.FA.Catalog.Application/SeedWork/Attributes/SanitizedAttribute.cs
@@ -1,4 +1,4 @@
-namespace Smart.FA.Catalog.Application;
+namespace Smart.FA.Catalog.Application.SeedWork.Attributes;
 
 /// <summary>
 /// Specifies that a property should be sanitized.
